Clang warnings dem (#1674)
Description
When compiling Lethe with Clang, multiple warning arise throughout the code. This removes the last remaining warningin the dem, fem-dem and rpt library. The majority of the warnings left were related to some virtual classes not having a virtual distructors and static_cast missing when using the type properties from the article. Lethe now compiles without any warning on modern Clang. This closes #1650

Testing
All of the tests should remain unchanged.

Documentation
Nothing to document

+++ include/dem/particle_particle_contact_force.h
@@ -663,9 +663,9 @@ private:
     // Get the reference of the effective properties according to the particle
     // types in vectors for easy-to-read equations.
     const unsigned int particle_one_type =
-      particle_one_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_one_properties[PropertiesIndex::type]);
     const unsigned int particle_two_type =
-      particle_two_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_two_properties[PropertiesIndex::type]);
     const unsigned int pair_index =
       vec_particle_type_index(particle_one_type, particle_two_type);
 
@@ -816,9 +816,9 @@ private:
                                      particle_two_properties);
 
     const unsigned int particle_one_type =
-      particle_one_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_one_properties[PropertiesIndex::type]);
     const unsigned int particle_two_type =
-      particle_two_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_two_properties[PropertiesIndex::type]);
     const unsigned int pair_index =
       vec_particle_type_index(particle_one_type, particle_two_type);
 
@@ -973,9 +973,9 @@ private:
                                      particle_two_properties);
 
     const unsigned int particle_one_type =
-      particle_one_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_one_properties[PropertiesIndex::type]);
     const unsigned int particle_two_type =
-      particle_two_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_two_properties[PropertiesIndex::type]);
     const unsigned int pair_index =
       vec_particle_type_index(particle_one_type, particle_two_type);
 
@@ -1113,9 +1113,9 @@ private:
                                      particle_two_properties);
 
     const unsigned int particle_one_type =
-      particle_one_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_one_properties[PropertiesIndex::type]);
     const unsigned int particle_two_type =
-      particle_two_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_two_properties[PropertiesIndex::type]);
     const unsigned int pair_index =
       vec_particle_type_index(particle_one_type, particle_two_type);
 
@@ -1241,9 +1241,9 @@ private:
                                      particle_two_properties);
 
     const unsigned int particle_one_type =
-      particle_one_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_one_properties[PropertiesIndex::type]);
     const unsigned int particle_two_type =
-      particle_two_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_two_properties[PropertiesIndex::type]);
     const unsigned int pair_index =
       vec_particle_type_index(particle_one_type, particle_two_type);
 
@@ -1403,9 +1403,9 @@ private:
                                      particle_two_properties);
 
     const unsigned int particle_one_type =
-      particle_one_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_one_properties[PropertiesIndex::type]);
     const unsigned int particle_two_type =
-      particle_two_properties[PropertiesIndex::type];
+      static_cast<unsigned int>(particle_two_properties[PropertiesIndex::type]);
     const unsigned int pair_index =
       vec_particle_type_index(particle_one_type, particle_two_type);
 
@@ -1910,9 +1910,11 @@ private:
             if (normal_overlap > 0)
               {
                 const unsigned int particle_one_type =
-                  particle_one_properties[PropertiesIndex::type];
+                  static_cast<unsigned int>(
+                    particle_one_properties[PropertiesIndex::type]);
                 const unsigned int particle_two_type =
-                  particle_two_properties[PropertiesIndex::type];
+                  static_cast<unsigned int>(
+                    particle_two_properties[PropertiesIndex::type]);
                 const unsigned int pair_index =
                   vec_particle_type_index(particle_one_type, particle_two_type);
                 const double temperature_one =
